rename DesktopTabs to RightPaneRoot
